Updates to Python 3.10
Updates the project to use Python 3.10.

This change upgrades the base Docker images in CircleCI,
the python version in pipenv, mypy, and setup.py to python 3.10.

Additionally, this change replaces pipenv-setup with a custom sync
script that updates the setup.py install_requires based on the
Pipfile.lock. This allows us to specify version ranges instead of
exact versions in setup.py for dependencies that are not daplug-core.